added database

+++ R/nm.R
@@ -96,31 +96,186 @@ nm <- function(cmd,psn_command,
       stop("couldn't infer unique run directory type.\nRerun with run_dir argument")
     message(paste("inferring run directory :",matched_run_dir))
     r$run_dir <- file.path(r$run_in,matched_run_dir)
-  }
+  } else r$run_dir <- run_dir
 
   if(missing(run_id)){
     matched_run_id <- get_run_id(r$ctl[1])
     if(length(matched_run_id)==0)
       stop("couldn't infer run id type.\nRerun with run_id argument")
     r$run_id <- matched_run_id
-  }
+  } else r$run_id <- run_id
 
   if(!file.exists(r$ctl)) stop(paste("cannot find model file:",r$ctl))
-  r$extn <- tools::file_ext(r$ctl)
   ## class for execute runs
   class(r) <- c(paste0("nm",r$type),class(r))
 
-  if(r$type %in% "execute"){
-    r$data_name <- data_name(r$ctl)
-    if(r$extn %in% "mod") r$lst <- gsub("^(.*\\.).*$","\\1lst",r$ctl) else
-      r$lst <- paste0(r$ctl,".lst")
-    r$psn.lst <- file.path(r$run_dir,"NM_run1","psn.lst")
-    r$psn.ext <- file.path(r$run_dir,"NM_run1","psn.ext")
-    r$psn.cov <- file.path(r$run_dir,"NM_run1","psn.cov")
-    r$psn.mod <- file.path(r$run_dir,"NM_run1","psn.mod")
-    r$param_info <- param_info(r$ctl)
-  }
+  r$input <- input_files(run_type = r$type,ctl_name = r$ctl)
+  r$output <- output_files(run_type = r$type,
+                           run_dir = r$run_dir,
+                           ctl_name = r$ctl)
+
+  if(r$type %in% "execute") r$param_info <- param_info(r$ctl)
+
+  ####
+
+  match_info <- nmdb_match_info(r)
+  matched_entry <- match_info$entry[match_info$match_type &
+                                      match_info$match_run_in &
+                                      match_info$match_ctl]
+
+  overlapped_output_entries <-
+    match_info$entry[match_info$overlap_outputs &
+                       !(match_info$entry %in% matched_entry)]
+  
+  if(length(overlapped_output_entries)>0)
+    stop("Outputs overlap with entries: ",
+         paste(overlapped_output_entries,collapse=","))
+
+  if(length(matched_entry)==0) {
+    nmdb_add_entry(r)
+  } else if(length(matched_entry)==1) {
+    query <- paste('DELETE FROM runs WHERE entry ==',matched_entry)
+    my_db <- DBI::dbConnect(RSQLite::SQLite(), ".runs.sqlite3")
+    DBI::dbExecute(my_db, query)
+    DBI::dbDisconnect(my_db)
+    message("Overwriting database entry: ",matched_entry)
+    nmdb_add_entry(r,matched_entry,silent=TRUE)
+  } else stop("Matched more than one database entry. Debug")
+
   return(r)
+}
+
+nmdb_match_info <- function(r){
+  match_rows <- c("type","run_dir")
+  if(!file.exists(".runs.sqlite3"))
+    return(data.frame(entry=numeric(),
+                      match_type=logical(),
+                      match_run_in=logical(),
+                      match_run_dir=logical(),
+                      match_ctl = logical(),
+                      overlap_outputs=logical()))
+
+  my_db <- DBI::dbConnect(RSQLite::SQLite(), ".runs.sqlite3")
+  tryCatch({
+    d <- DBI::dbGetQuery(my_db, 'SELECT * FROM runs')
+    DBI::dbDisconnect(my_db)
+    dproposal <- nmdb_make_db_row(r)
+
+    ## objectives
+    ## if a run is of the same type, run_dir, and outputfiles,
+    ans <- list()
+
+    ans$entry <- d$entry
+    ans$match_type <- d$type %in% r$type
+    ans$match_run_dir <- d$run_dir %in% r$run_dir
+    ans$match_run_in <- d$run_in %in% r$run_in
+    ans$match_ctl <- d$ctl %in% r$ctl
+
+    dependent_run_dir <- sapply(d$run_dir,function(i)grepl(paste0("^",i),r$run_dir))
+    ans$dependent_run_dir <- dependent_run_dir & !ans$match_run_dir
+
+    overlap_outputs <- sapply(strsplit(d$output_files,","),function(out_filesi){
+      length(intersect(out_filesi,unlist(r$output)))>0
+    })
+    ans$overlap_outputs <- overlap_outputs
+
+    ans <- as.data.frame(ans)
+    return(ans)
+  },
+  error=function(e) {
+    if(DBI::dbIsValid(my_db)) DBI::dbDisconnect(my_db)
+    stop(e)
+  })
+}
+
+nmdb_make_db_row <- function(r){
+  if(r$type %in% "execute") r$input$data <- NULL
+  data.frame(object=I(list(serialize(r,NULL))),
+             type=r$type,
+             run_in=r$run_in,
+             run_dir=r$run_dir,
+             ctl = r$ctl,
+             cmd = r$cmd,
+             output_files=paste(unlist(r$output),collapse=","),
+             input_files=paste(unlist(r$input),collapse=","))
+}
+
+nmdb_printable_db <- function(d){
+  d$object <- NULL
+  d$output_files <- NULL
+  d$input_files <- lapply(strsplit(d$input_files,","),function(char_vec){
+    paste(basename(char_vec),collapse=",")
+  })
+  d$ctl <- NULL
+  d
+}
+
+nmdb_add_entry <- function(r,entry=NULL,silent=FALSE){
+  new_db_flag <- !file.exists(".runs.sqlite3")
+  dnew <- nmdb_make_db_row(r)
+  my_db <- DBI::dbConnect(RSQLite::SQLite(), ".runs.sqlite3")
+  tryCatch({
+    if(new_db_flag) {
+      dempty <- dnew[c(),]
+      dempty <- cbind(data.frame(entry=numeric()),dempty)
+      DBI::dbWriteTable(my_db, "runs", dempty)
+    }
+    d <- DBI::dbGetQuery(my_db, 'SELECT * FROM runs')
+
+    if(nrow(d)==0) new_entry <- 1 else
+      if(!is.null(entry)) new_entry <- entry else
+        new_entry <- max(d$entry) + 1
+
+    dnew <- cbind(data.frame(entry=new_entry),dnew)
+
+    if(!silent) message("Creating database entry: ",new_entry)
+    DBI::dbWriteTable(my_db, "runs", dnew, append=TRUE)
+    DBI::dbDisconnect(my_db)
+    invisible()
+  },
+  error=function(e){
+    if(DBI::dbIsValid(my_db)) DBI::dbDisconnect(my_db)
+    if(new_db_flag) unlink(".runs.sqlite3",force = TRUE)
+    stop(e)
+  })
+}
+
+nmdb_get <- function(readable=FALSE){
+  my_db <- DBI::dbConnect(RSQLite::SQLite(), ".runs.sqlite3")
+  tryCatch({
+    d <- DBI::dbGetQuery(my_db, 'SELECT * FROM runs')
+    if(readable) d <- nmdb_printable_db(d)
+    DBI::dbDisconnect(my_db)
+    d <- d[order(d$entry),]
+    return(d)
+  },
+  error=function(e) {
+    if(DBI::dbIsValid(my_db)) DBI::dbDisconnect(my_db)
+    stop(e)
+  })
+}
+
+nmdb_display <- function(...) nmdb_get(readable = TRUE,...)
+
+#' Extract nm object from data.base
+#' @param entry numeric. Database ID of nm object
+#' @export
+
+nmdb_extract <- function(entry){
+  my_db <- DBI::dbConnect(RSQLite::SQLite(), ".runs.sqlite3")
+  tryCatch({
+    d <- DBI::dbGetQuery(my_db, 'SELECT * FROM runs')
+    DBI::dbDisconnect(my_db)
+    d <- d[d$entry==entry, ]
+    if(nrow(d)>1) stop("More than one entry found")
+    if(nrow(d)==0) stop("No entry found")
+    r <- unserialize(d$object[[1]])
+    return(r)
+  },
+  error=function(e){
+    if(DBI::dbIsValid(my_db)) DBI::dbDisconnect(my_db)
+    e
+  })
 }
 
 #' @export
@@ -140,32 +295,19 @@ get_run_id <- function(ctl_name){
   gsub2(match,"\\1",ctl_name)
 }
 
-#' Load nm object into memory from script
-#'
-# nm_load <- function(script){
-#   script <- tidyproject::locate_file(script,getOption("scripts.dir"))
-#   s <- readLines(script)
-#   s <- rem_comment(s,"#")  ## this is simplistic but good enough for now
-#   ## wont ignore a # in a string for example or an escaped #
-#   s <- s[grepl("\\s*nm\\(\"",s)] ## detect nm(...) rows
-#   if(length(s)==0) stop("no nm() lines detected")
-#   message("running...")
-#   message(paste(s,collapse="\n"))
-#   eval(parse(text = s),envir = .GlobalEnv)
-# }
-
-
 #' Run NONMEM
 #' @param ... objects of class nm
 #' @param overwrite logical. Should run directory be overwritten (default=FALSE)
+#' @param delete_dir logical. NA (default - directory will be deleted if no dependencies exists)
+#' TRUE or FALSE. Should run_dir be deleted.
 #' @export
-run <- function(...,overwrite=FALSE){
+run <- function(...,overwrite=FALSE,delete_dir=c(NA,TRUE,FALSE)){
   rl <- list(...)
   lapply(rl,function(r){
     ## if directory exists, and if it's definately a directory stop
     if(file.exists(r$run_dir) & !overwrite)if(file.info(r$run_dir)$isdir %in% TRUE) stop("run already exists. To rerun select overwrite=TRUE")
     if(!is.null(getOption("kill_run"))){
-      getOption("kill_run")(r)
+      getOption("kill_run")(r,delete_dir=match.arg(delete_dir))
     }
     clean_run(r)
     system_nm(cmd = r$cmd,dir = r$run_in)
@@ -181,14 +323,24 @@ nm_tran.nm <- function(x){
 #' clean_run files
 #'
 #' @param r character or numeric. run identifier
+#' @param delete_dir logical. NA (default), TRUE or FALSE. Should run_dir be deleted.
 #' @export
-clean_run <- function(r){
+clean_run <- function(r,delete_dir=c(NA,TRUE,FALSE)){
   ## assumes ctrl file is run[run_id].mod and -dir=[run_id] was used
-  unlink(out_files0(r$ctl))
-  unlink(r$run_dir,recursive=TRUE) ## delete run directory
+  unlink(ctl_out_files(r$ctl))
+  delete_dir <- match.arg(delete_dir)
+  if(is.na(delete_dir)){
+    match_info <- nmdb_match_info(r)
+    matched_entry <- match_info$entry[match_info$dependent_run_dir]
+    if(length(matched_entry)>0) {
+      stop("Dependent run(s) found: ",paste(matched_entry,collapse=","),
+           "\nRerun with delete_dir=TRUE or delete_dir=FALSE or remove dependent run")
+    }
+  }
+  if(TRUE %in% delete_dir) unlink(r$run_dir,recursive=TRUE) ## delete run directory
 }
 
-out_files0 <- function(ctrl.file){ ## will get vector of $TABLE file names from control file.
+ctl_out_files <- function(ctrl.file){ ## will get vector of $TABLE file names from control file.
   if(!file.exists(ctrl.file)) stop(paste(ctrl.file, "doesn't exist"))
   dir0 <- dir(dirname(ctrl.file))
 
@@ -207,4 +359,31 @@ out_files0 <- function(ctrl.file){ ## will get vector of $TABLE file names from 
   out.files <- c(table.files,out.files)
   out.files
 }
+
+input_files <- function(run_type,ctl_name){
+  r <- list()
+  if(run_type %in% "execute"){
+    r$ctl <- ctl_name
+    r$data_name <- data_name(ctl_name)
+  }
+  return(r)
+}
+
+output_files <- function(run_type,run_dir,ctl_name){
+  r <- list()
+  if(run_type %in% "execute"){
+    extn <- tools::file_ext(ctl_name)
+    if(extn %in% "mod") r$lst <- gsub("^(.*\\.).*$","\\1lst",ctl_name) else
+      r$lst <- paste0(ctl_name,".lst")
+    r$psn.mod <- file.path(run_dir,"NM_run1","psn.mod")
+    r$psn.lst <- file.path(run_dir,"NM_run1","psn.lst")
+    r$psn.ext <- file.path(run_dir,"NM_run1","psn.ext")
+    r$psn.cov <- file.path(run_dir,"NM_run1","psn.cov")
+    r$psn.cor <- file.path(run_dir,"NM_run1","psn.cor")
+    r$psn.xml <- file.path(run_dir,"NM_run1","psn.xml")
+    r$ctl_out_files <- file.path(run_dir,ctl_out_files(ctl_name))
+  }
+  return(r)
+}
+
